Add type snapshots to module graph tests
