+++ crates/biome_js_type_info/Cargo.toml
@@ -22,4 +22,5 @@ biome_rowan               = { workspace = true }
 [dev-dependencies]
 biome_js_formatter = { workspace = true }
 biome_js_parser    = { workspace = true }
+biome_test_utils   = { workspace = true }
 insta              = { workspace = true }
