+++ crates/biome_test_utils/Cargo.toml
@@ -22,6 +22,7 @@ biome_diagnostics    = { workspace = true }
 biome_formatter      = { workspace = true }
 biome_fs             = { workspace = true }
 biome_js_parser      = { workspace = true }
+biome_js_type_info   = { workspace = true }
 biome_json_parser    = { workspace = true }
 biome_module_graph   = { workspace = true }
 biome_package        = { workspace = true }
