+++ crates/biome_test_utils/src/lib.rs
@@ -9,6 +9,7 @@ use biome_diagnostics::termcolor::Buffer;
 use biome_diagnostics::{DiagnosticExt, Error, PrintDiagnostic};
 use biome_fs::{BiomePath, FileSystem, OsFileSystem};
 use biome_js_parser::{AnyJsRoot, JsFileSource, JsParserOptions};
+use biome_js_type_info::{NUM_PREDEFINED_TYPES, TypeResolver, TypeResolverLevel};
 use biome_json_parser::{JsonParserOptions, ParseDiagnostic};
 use biome_module_graph::ModuleGraph;
 use biome_package::PackageJson;
@@ -282,6 +283,32 @@ fn markup_to_string(markup: biome_console::Markup) -> String {
     fmt.write_markup(markup).unwrap();
 
     String::from_utf8(buffer).unwrap()
+}
+
+pub fn dump_registered_types(content: &mut String, resolver: &dyn TypeResolver) {
+    let mut registered_types = String::new();
+    let mut resolver = Some(resolver);
+    while let Some(current_resolver) = resolver {
+        for (i, ty) in current_resolver.registered_types().iter().enumerate() {
+            let level = current_resolver.level();
+            let id = if level == TypeResolverLevel::Global {
+                i + NUM_PREDEFINED_TYPES
+            } else {
+                i
+            };
+            registered_types.push_str(&format!("\n{level:?} TypeId({id}) => {ty}\n"));
+        }
+
+        resolver = current_resolver.fallback_resolver();
+    }
+
+    if !registered_types.is_empty() {
+        content.push_str("## Registered types\n\n");
+
+        content.push_str("```\n");
+        content.push_str(&registered_types);
+        content.push_str("\n```\n");
+    }
 }
 
 // Check that all red / green nodes have correctly been released on exit
